Add trace-search package
Added the trace-search package via git+https since we haven't published to NPM yet.
This includes an OAuth token for our GitHub bot account.

This would usually be v bad to commit, but the bot is limited to
read-only, and the search repo will eventually be public anyway.
We'll deactivate the token when the repos go public.