First shot at injecting the analytics code
The requests are working and returning 200, but aren't appearing in Matomo.
Thinking that this is something with the referrer/origin/URL being ignored in the reports.
The same request shows up when sent from Postman, however, so slightly confused.

EDIT: Solved it. Browser was sending the DNT header and Matomo was configured to listen to
obey it. Turned DNT support off for the sake of the demo. Links were helpful:

https://developer.matomo.org/api-reference/tracking-api#debugging-the-tracker

https://analytics.tracedigital.tk/index.php?module=PrivacyManager&action=usersOptOut